refactor(name): 온보딩 컴포넌트 폴더명 변경
