+++ src/components/onboading/Progress2.jsx
