+++ src/components/onboading/Progress6.jsx
