+++ src/pages/onboading/ProfileLanguagePage.js
@@ -7,7 +7,7 @@ import { CustomTheme } from '@styles/CustomTheme.js';
 import { useOnboarding } from 'src/states/OnboardingContext.js';
 
 import ArrowRight32 from '@components/Icon32/ArrowRight32';
-import Progress5 from '@components/OnboadingCompo/Progress5';
+import Progress5 from '@components/onboading/Progress5';
 import Checkbox from '@components/common/Checkbox';
 import ApplyButton from '@components/common/ApplyButton';
